# Optimize add_items for strings.