Summary:
Optimizations 
1. Avoid calling setNoCopy and directly set output in result vector data
2. add Fast path for null free input. 
3. add fast path for flat input. 


before:
```
array_concat_VARCHAR_10##2arg                             368.39ms      2.71
array_concat_VARCHAR_10##3arg                             476.22ms      2.10
array_concat_VARCHAR_10##4arg                             580.82ms      1.72
array_concat_VARCHAR_20##2arg                             650.82ms      1.54
array_concat_VARCHAR_20##3arg                             848.01ms      1.18
array_concat_VARCHAR_20##4arg                                1.10s   911.34m
array_concat_VARCHAR_40##2arg                                1.03s   972.17m
array_concat_VARCHAR_40##3arg                                1.54s   651.30m
array_concat_VARCHAR_40##4arg                                1.98s   503.94m
array_concat_VARCHAR_5##2arg                              154.41ms      6.48
array_concat_VARCHAR_5##3arg                              209.87ms      4.76
array_concat_VARCHAR_5##4arg                              272.49ms      3.67
```

after:
```
array_concat_VARCHAR_10##2arg                             192.25ms      5.20
array_concat_VARCHAR_10##3arg                             229.85ms      4.35
array_concat_VARCHAR_10##4arg                             292.27ms      3.42
array_concat_VARCHAR_20##2arg                             256.81ms      3.89
array_concat_VARCHAR_20##3arg                             334.52ms      2.99
array_concat_VARCHAR_20##4arg                             462.94ms      2.16
array_concat_VARCHAR_40##2arg                             348.40ms      2.87
array_concat_VARCHAR_40##3arg                             459.01ms      2.18
array_concat_VARCHAR_40##4arg                             696.25ms      1.44
array_concat_VARCHAR_5##2arg                              104.52ms      9.57
array_concat_VARCHAR_5##3arg                              140.64ms      7.11
array_concat_VARCHAR_5##4arg                              192.35ms      5.20
```

Differential Revision: D52450883